moved load function into separate file
